requirements.txt
langchain
openai
fastapi
python-dotenv
python-multipart
pyjwt 
passlib
python-jose[cryptography]